% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average_piks_list}
\alias{average_piks_list}
\title{Average of priors: p0k, pak and pck from list (memory intensive)}
\usage{
average_piks_list(params, nsnps, rg_vec, nits, thin, rg = FALSE)
}
\arguments{
\item{params}{Vector of parameters: α, β and γ}

\item{nsnps}{number of snps}

\item{rg_vec}{Vector of the covariate}

\item{nits}{Number of iterations run in mcmc}

\item{thin}{thinning}

\item{rg}{logical: was the covariate inflormation  used? default: False}
}
\value{
average pik matrix of priors: p0k, pak and pck
}
\description{
Average of priors: p0k, pak and pck from list (memory intensive)
}
