% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average_posterior_prob}
\alias{average_posterior_prob}
\title{Average of posterior probabilities: Hn, Ha and Hc}
\usage{
average_posterior_prob(params, lbf_mat, nsnps, rg_vec, nits, thin, rg = FALSE)
}
\arguments{
\item{params}{Vector of parameters: α, β and γ}

\item{lbf_mat}{matrix of log bayes factors: lbfak and lbfck}

\item{nsnps}{number of snps}

\item{rg_vec}{Vector of the covariate}

\item{nits}{Number of iterations run in mcmc}

\item{thin}{thinning}

\item{rg}{logical: was the covariate inflormation  used? default: False}
}
\value{
matrix with average of all the posterior probabilities: Hn, Ha and Hc
}
\description{
Average of posterior probabilities: Hn, Ha and Hc
}
