% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{causal.priors}
\alias{causal.priors}
\title{causal.priors}
\usage{
causal.priors(p1, p12)
}
\arguments{
\item{p1}{prior probability a SNP is associated with trait 1}

\item{p12}{prior probability a SNP is associated with both traits}
}
\value{
priors at the causal variant
}
\description{
Estimate priors at the known causal variant of trait 1
}
\author{
Ichcha Manipur
}
