% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrop_hier_priors.R
\name{combine.bf.kc.hier}
\alias{combine.bf.kc.hier}
\title{combine.bf.kc.hier}
\usage{
combine.bf.kc.hier(pik_vec, lbfk_vec)
}
\arguments{
\item{pik_vec}{named vector (pnk, pak, pck)}

\item{lbfk_vec}{named log bayes factor vector (lBF.Ha, lBF.Hc)}
}
\value{
named numeric vector of posterior probabilities and bayes factors
}
\description{
Calculate posterior probabilities from priors, given logABFs for each SNP
and priors inferred from the hierarchical model
}
\author{
Ichcha Manipur
}
