% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{combine.bf.kc}
\alias{combine.bf.kc}
\title{combine.bf.kc}
\usage{
combine.bf.kc(labf, pn, pa, pc, causalpos1)
}
\arguments{
\item{labf}{log approximate bayes factors}

\item{pn}{prior probability a SNP other than the causal variant is associated with trait 1}

\item{pa}{prior probability a SNP other than the causal variant is associated with trait 2}

\item{pc}{prior probability the causal SNP of trait 1 is associated with both traits}

\item{causalpos1}{Position of trait1 causal SNP}
}
\value{
named numeric vector of posterior probabilities and bayes factors
}
\description{
Internal function, calculate posterior probabilities for configurations, given logABFs for each SNP and prior probs and position of known causal variant of trait 1
}
\author{
Ichcha Manipur
}
