% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{combine.bf}
\alias{combine.bf}
\title{combine.bf}
\usage{
combine.bf(lBF_df, pn, pa, pc)
}
\arguments{
\item{pn}{prior probability that none of the SNPs/variants in the region are associated with the query trait}

\item{pa}{prior probability that a non-query variant is causally associated with the query trait}

\item{pc}{prior probability that the query variant is causally associated with the query trait}

\item{lBF_mat}{log bayes factors}
}
\value{
named numeric vector of posterior probabilities and bayes factors
}
\description{
Internal function, calculate posterior probabilities for all the configurations
}
\author{
Ichcha Manipur
}
