% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{cophe.impute}
\alias{cophe.impute}
\title{cophe.impute
Imputation of untyped SNPs with summary statistics using the SSIMP algorithm}
\usage{
cophe.impute(dataset, LD, MAF, nsample, r2.thr = 1)
}
\arguments{
\item{dataset}{dataset conforming to coloc, Use coloc::check_dataset() for}

\item{LD}{LD matrix including all SNPs to be imputed}

\item{MAF}{Minimum allele frequency}

\item{nsample}{number of samples from which the LD matrix is derived}

\item{r2.thr}{tags are selected with max r2 <= r2.thr}
}
\value{
Imputed dataset
}
\description{
cophe.impute
Imputation of untyped SNPs with summary statistics using the SSIMP algorithm
}
