% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{cophe.prepare.dat.single}
\alias{cophe.prepare.dat.single}
\title{Prepare data for cophe.single}
\usage{
cophe.prepare.dat.single(dataset, querysnpid, MAF = NULL)
}
\arguments{
\item{dataset}{a list with specifically named elements defining the query trait dataset
to be analysed.}

\item{querysnpid}{Id of the query variant, (id in dataset$snp)}
}
\value{
named list containing the per snp BFs (df) and position of the query variant (querypos)
}
\description{
Prepare data for cophe.single
}
\author{
Ichcha Manipur
}
\keyword{internal}
