% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarsusie.R
\name{cophe.prepare.dat.susie}
\alias{cophe.prepare.dat.susie}
\title{Prepare data for \code{cophe.susie}}
\usage{
cophe.prepare.dat.susie(dataset, querysnpid, susie.args)
}
\arguments{
\item{dataset}{a list with specifically named elements defining the query trait dataset
to be analysed.}

\item{querysnpid}{Id of the query variant, (id in dataset$snp)}

\item{susie.args}{a named list of additional arguments to be passed to
\link[coloc]{runsusie}}
}
\value{
a list with the output of running susie
}
\description{
Prepare data for \code{cophe.susie}
}
\seealso{
\code{\link{cophe.susie}}
}
\keyword{internal}
