% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{cophe.single.lbf}
\alias{cophe.single.lbf}
\title{cophe.single.lbf}
\usage{
cophe.single.lbf(dataset, querysnpid, querytrait, MAF = NULL)
}
\arguments{
\item{dataset}{a list with specifically named elements defining the query trait dataset
to be analysed.}

\item{querysnpid}{Id of the query variant, (id in dataset$snp)}

\item{querytrait}{Query trait name}

\item{MAF}{Minor allele frequency vector}
}
\value{
data frame with log bayes factors for Hn and Ha hypotheses
}
\description{
cophe.single.lbf
}
\author{
Ichcha Manipur
}
