% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{cophe.single}
\alias{cophe.single}
\title{Fully Bayesian cophe analysis using Bayes Factors}
\usage{
cophe.single(
  dataset,
  causal.snpid,
  MAF = NULL,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05,
  p2a = NULL,
  p12c = NULL,
  forhier = F
)
}
\arguments{
\item{dataset}{a list with specifically named elements defining the dataset
to be analysed.}

\item{MAF}{Minor allele frequency vector to be used for  dataset, a sHarthand for supplying the same vector as parts of both datasets}

\item{p1}{prior probability a SNP is associated with trait 1, default 1e-4}

\item{p2}{prior probability a SNP is associated with trait 2, default 1e-4}

\item{p12}{prior probability a SNP is associated with both traits, default 1e-5}

\item{p2a}{prior probability a SNP other than the causal variant is associated with trait 2, default 1e-4}

\item{p12c}{prior probability the causal SNP of trait 1 is associated with both traits, default 1e-5}

\item{causalpos1}{Position of trait1 causal SNP}
}
\value{
a list of two \code{data.frame}s:
\itemize{
\item summary is a vector giving the number of SNPs analysed, and the posterior probabilities of Hn (no shared causal variant), Ha (two distinct causal variants) and Hc (one common causal variant)
\item results is an annotated version of the input data containing log Approximate Bayes Factors and intermediate calculations, and the posterior probability SNP.PP.Hc of the SNP being causal for the shared signal \emph{if} Hc is true. This is only relevant if the posterior support for Hc in summary is convincing.
}
}
\description{
Bayesian cophe analysis for single causal variant assumption
Not to be used
}
\details{
This function calculates posterior probabilities of different
causal variant configurations under the assumption of a single
causal variant for each trait.

If regression coefficients and variances are available, it
calculates Bayes factors for association at each SNP.  If only p
values are available, it uses an approximation that depends on the
SNP's MAF and ignores any uncertainty in imputation.  Regression
coefficients sHauld be used if available.
}
\author{
Ichcha Manipur
}
