% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarsusie.R
\name{cophe.susie.lbf}
\alias{cophe.susie.lbf}
\title{cophe.susie.lbf}
\usage{
cophe.susie.lbf(
  dataset,
  querysnpid,
  querytrait,
  switch = TRUE,
  susie.args = list(),
  MAF = NULL
)
}
\arguments{
\item{dataset}{a list with specifically named elements defining the query trait dataset
to be analysed.}

\item{querysnpid}{Id of the query variant, (id in dataset$snp)}

\item{querytrait}{Query trait name}

\item{switch}{Set switch=TRUE to obtain single BF when credible sets not found with SuSIE}

\item{susie.args}{a named list of additional arguments to be passed to
\link{runsusie}}

\item{MAF}{Minor allele frequency vector}
}
\value{
data frame with log bayes factors for Hn and Ha hypotheses
}
\description{
cophe.susie.lbf
}
\author{
Ichcha Manipur
}
