% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{cophe_heatmap}
\alias{cophe_heatmap}
\title{Heatmap of multi-trait cophescan results}
\usage{
cophe_heatmap(
  multi.dat,
  thresh_Hc = 0.5,
  thresh_Ha = 0.5,
  cluster_columns = F,
  cluster_rows = F,
  cell_width = 20,
  cell_height = 20,
  ...
)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait}

\item{thresh}{Traits with posterior probabilities above this threshold for Hc and Ha will be displayed}
}
\value{
heatmap of posterior probabilities of the phentypes above the set threshold
}
\description{
Heatmap of multi-trait cophescan results
}
