% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{cophe_multi_trait_data}
\alias{cophe_multi_trait_data}
\title{Simulated multi-trait data}
\format{
list of coloc structred datasets for 24 traits (cophe_multi_trait_data$summ_stat), LD matrix (cophe_multi_trait_data$LD) and the id of the query snp  (cophe_multi_trait_data$querysnpid). #' The trait dataset  are simulated summary statistics (1000 SNPs) for 10 Hn, 10 Ha and 10 Hc.
}
\usage{
data(cophe_multi_trait_data)
}
\description{
Simulated multi-trait data
}
\keyword{datasets}
