% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{cophe_plot}
\alias{cophe_plot}
\title{cophe_plot}
\usage{
cophe_plot(multi.dat, thresh = 0.5, addplot = NULL, adddat = NULL)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait}

\item{thresh}{Traits with posterior probabilities above this threshold for Hc and Ha will be labelled}
}
\value{
cophescan plots of Ha and Hc
}
\description{
cophe_plot
}
