% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_posterior_prob}
\alias{get_posterior_prob}
\title{Calculation of the posterior prob of Hn, Ha and Hc}
\usage{
get_posterior_prob(params, lbf_mat, nsnps, rg_vec, rg = FALSE)
}
\arguments{
\item{lbf_mat}{matrix of log bayes factors: lbfak and lbfck}

\item{nsnps}{number of snps}

\item{rg_vec}{Vector of the covariate}

\item{rg}{logical: should the covariate inflormation be used? default: False}

\item{pars}{Vector of parameters: α, β and γ}
}
\value{
posterior prob of Hn, Ha and Hc
}
\description{
Calculation of the posterior prob of Hn, Ha and Hc
}
