% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophe_hyp_predict.R
\name{Hc.cutoff.fdr}
\alias{Hc.cutoff.fdr}
\title{Estimate the Hc.cutoff for the required FDR}
\usage{
Hc.cutoff.fdr(ppHc, fdr = 0.05, return_plot = TRUE)
}
\arguments{
\item{ppHc}{a vector containing the PP.Hc (the posterior probability of causal association) of all tests}

\item{fdr}{FDR default: 0.05}

\item{return_plot}{default: TRUE, plot the fdr estimated at the different Hc.cutoff}
}
\value{
the Hc.cutoff value for the specified FDR, if return_plot is True returns a plot showing the FDR calculated at different Hc thresholds
}
\description{
Estimate the Hc.cutoff for the required FDR
}
