% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{hypothesis.priors}
\alias{hypothesis.priors}
\title{hypothesis.priors}
\usage{
hypothesis.priors(pa, pc, nsnps)
}
\arguments{
\item{pa}{prior probability a SNP other than the causal variant is associated with trait 2, default 1e-4}

\item{pc}{prior probability the causal SNP of trait 1 is associated with both traits, default 1e-5}

\item{nsnps}{number of SNPs}
}
\value{
hypotheses priors
}
\description{
hypothesis.priors
}
\author{
Ichcha Manipur
}
