% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logd_alpha}
\alias{logd_alpha}
\title{dnorm for alpha}
\usage{
logd_alpha(a, mean = -10, sd = 0.5, log = TRUE)
}
\arguments{
\item{alpha_mean}{prior for the mean of  alpha}

\item{alpha_sd}{prior for the standard deviation of  alpha}
}
\description{
dnorm for alpha
}
