% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logd_beta}
\alias{logd_beta}
\title{dgamma for beta}
\usage{
logd_beta(b, beta_shape = 2, beta_scale = 2)
}
\arguments{
\item{b}{current beta}

\item{beta_shape}{prior for the shape (gamma distibution) of beta}

\item{beta_scale}{prior for the scale of beta}
}
\value{
log dgamma
}
\description{
dgamma for beta
}
