% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logd_gamma}
\alias{logd_gamma}
\title{dgamma for gamma}
\usage{
logd_gamma(g, gamma_shape = 2, gamma_scale = 2)
}
\arguments{
\item{g}{current gamma}

\item{gamma_shape}{prior for the shape (gamma distibution) of gamma}

\item{gamma_scale}{prior for the scale of gamma}
}
\value{
log dgamma
}
\description{
dgamma for gamma
}
