% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik}
\alias{loglik}
\title{Log likelihood calculation}
\usage{
loglik(params, lbf_mat, nsnps, covar_vec, covar = FALSE)
}
\arguments{
\item{params}{Vector of parameters: \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma}}

\item{lbf_mat}{matrix of log bayes factors: lBF.Ha and lBF.Hc}

\item{nsnps}{number of snps}

\item{covar_vec}{Vector of the covariate}

\item{covar}{logical: should the covariate information be used? default: False}
}
\value{
logpost log of the posteriors
}
\description{
Log likelihood calculation
}
