% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logpriors}
\alias{logpriors}
\title{Proposal distribution}
\usage{
logpriors(
  params,
  rg = FALSE,
  alpha_mean = -10,
  alpha_sd = 0.5,
  beta_shape = 2,
  beta_scale = 2,
  gamma_shape = 2,
  gamma_scale = 2
)
}
\description{
Proposal distribution
}
