% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{multi_plot}
\alias{multi_plot}
\title{multi_plot}
\usage{
multi_plot(
  multi.dat,
  thresh = 0.5,
  pltheatmap = F,
  cophe.plots = T,
  addplot = NULL,
  adddat = NULL
)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait}

\item{thresh}{Traits with posterior probabilities above the threshold for Hc and Ha  will be labelled}

\item{cophe.plots}{Plot cophe.plots. Similar to PheWAS plots with Hc and Ha on the y axis}

\item{heatmap}{Plot heatmap}
}
\value{
plot list
}
\description{
multi_plot
}
