% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pars2pik}
\alias{pars2pik}
\title{Conversion of parameters α, β and γ to p0k, pak and pck}
\usage{
pars2pik(params, nsnps, rg_vec, rg = FALSE)
}
\arguments{
\item{params}{Vector of parameters: α, β and γ}

\item{nsnps}{number of snps}

\item{rg_vec}{Vector of the covariate}

\item{rg}{logical: should the covariate information be used? default: False}
}
\value{
pik matrix of priors: p0k, pak and pck
}
\description{
Conversion of parameters α, β and γ to p0k, pak and pck
}
