% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piks}
\alias{piks}
\title{piks}
\usage{
piks(params, nsnps, rg_vec, rg = FALSE)
}
\arguments{
\item{params}{Matrix of hyperprior distributions}

\item{nsnps}{Vector of number of snps}

\item{rg_vec}{Vector of genetmic correlation}

\item{rg}{Include genetic correlation default:FALSE}
}
\value{
A list of prior probabilities of p0k, pak and pck
}
\description{
Calculate the prior probabilities p0k, pak and pck from posteriors of alpha, beta and gamma
}
