% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{plot_cophe_ternary}
\alias{plot_cophe_ternary}
\title{Ternary plots of multi-trait cophescan output}
\usage{
plot_cophe_ternary(
  multi.dat,
  querysnpid,
  traits.dat = NULL,
  plot_pval = F,
  thresh_Hc = 0.5,
  thresh_Ha = 0.5
)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait}

\item{querysnpid}{the id of the causal variant as present in trait.dat$snp, , plotted in red}

\item{traits.dat}{list of multi-trait coloc structured datasets}

\item{plot_pval}{Map p_value of Queried variant}

\item{thresh_Hc}{Hc threshold to be displayed}

\item{thresh_Ha}{Ha threshold to be displayed}
}
\value{
ternary plot of the posterior probabilities of Ha, Hc and Hn
}
\description{
Ternary plots of multi-trait cophescan output
}
