% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{plot_trait_manhat}
\alias{plot_trait_manhat}
\title{Plot region Manhattan for a trait highlighting the queried variant}
\usage{
plot_trait_manhat(trait.dat, causal.snpid)
}
\arguments{
\item{trait.dat}{dataset used as input for running cophescan}

\item{causal.snpid}{the id of the causal variant as present in trait.dat$snp}
}
\value{
region manhattan plot
}
\description{
Plot region Manhattan for a trait highlighting the queried variant
}
