% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{post_prob}
\alias{post_prob}
\title{post_prob}
\usage{
post_prob(params, lbf_mat, nsnps, rg_vec, rg = FALSE)
}
\arguments{
\item{params}{Matrix of hyperprior distributions}

\item{lbf_mat}{Matrix of bayes factors}

\item{nsnps}{Vector of number of snps}

\item{rg_vec}{Vector of genetic correlation}

\item{rg}{Include genetic correlation default:FALSE}
}
\value{
A list of posterior probabilities of Hn, Ha and Hc
}
\description{
Calculate the posterior probabilities of Hn, Ha and Hc from the output of piks
}
