% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propose}
\alias{propose}
\title{Proposal distribution}
\usage{
propose(params, propsd = 0.5)
}
\arguments{
\item{params}{Vector of parameters: α, β and γ}

\item{propsd}{Standard deviation for the proposal}
}
\value{
vector : proposal
}
\description{
Proposal distribution
}
