% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrop_hier_priors.R
\name{run_metrop_priors}
\alias{run_metrop_priors}
\title{run_metrop_priors}
\usage{
run_metrop_priors(
  lbf_mat,
  rg = FALSE,
  rg_vec = NULL,
  nits = 10000,
  thin = 1000
)
}
\arguments{
\item{rg}{whether to run rg}

\item{rg_vec}{vector of genetic correlations}

\item{nits}{number of iterations}

\item{thin}{burnin}

\item{multi.dat}{matrix of bf values}
}
\value{
list containing posterior of the parameters
}
\description{
run_metrop_priors
}
