% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_alpha}
\alias{sample_alpha}
\title{sample alpha}
\usage{
sample_alpha(alpha_mean = -10, alpha_sd = 0.5)
}
\arguments{
\item{alpha_mean}{prior for the mean of  alpha}

\item{alpha_sd}{prior for the standard deviation of  alpha}
}
\value{
sample from rnorm for \eqn{\alpha}
}
\description{
sample alpha
}
