% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_beta}
\alias{sample_beta}
\title{sample beta}
\usage{
sample_beta(beta_shape = 2, beta_scale = 2)
}
\arguments{
\item{beta_shape}{prior for the shape (gamma distibution) of beta}

\item{beta_scale}{prior for the scale of beta}
}
\value{
sample from rgamma for \eqn{\beta}
}
\description{
sample beta
}
