% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{summary.cophe}
\alias{summary.cophe}
\title{Title}
\usage{
\method{summary}{cophe}(cophe.res)
}
\arguments{
\item{cophe.res}{Result from either cophe.susie or cophe.single}
}
\value{
log bayes and posterior probabilities
}
\description{
Title
}
