% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average_posterior_prob}
\alias{average_posterior_prob}
\title{Average of posterior probabilities: Hn, Ha and Hc}
\usage{
average_posterior_prob(
  params,
  lbf_mat,
  nsnps,
  covar_vec,
  nits,
  thin,
  covar = FALSE
)
}
\arguments{
\item{params}{Vector of parameters: \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma}}

\item{lbf_mat}{matrix of log bayes factors: lBF.Ha and lBF.Hc}

\item{nsnps}{number of snps}

\item{covar_vec}{Vector of the covariate}

\item{nits}{Number of iterations run in mcmc}

\item{thin}{thinning}

\item{covar}{logical: was the covariate inflormation  used? default: False}
}
\value{
matrix with average of all the posterior probabilities: Hn, Ha and Hc
}
\description{
Average of posterior probabilities: Hn, Ha and Hc
}
