% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{combine.bf.kc}
\alias{combine.bf.kc}
\title{combine.bf.kc}
\usage{
combine.bf.kc(l2, p2a, p12c, causalpos1)
}
\arguments{
\item{l2}{df$lABF.df2}

\item{p2a}{prior probability a SNP other than the causal variant is associated with trait 2, default 1e-4}

\item{p12c}{prior probability the causal SNP of trait 1 is associated with both traits, default 1e-5}

\item{causalpos1}{Position of trait1 causal SNP}
}
\value{
named numeric vector of posterior probabilities and bayes factors
}
\description{
Internal function, calculate posterior probabilities for configurations, given logABFs for each SNP and prior probs and position of known causal variant of trait 1
}
\author{
Ichcha Manipur
}
