% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarsusie.R
\name{cophe.bf_bf}
\alias{cophe.bf_bf}
\title{extract data through Bayes factors}
\usage{
cophe.bf_bf(bf2, causal.snpid, pn, pa, pc)
}
\arguments{
\item{bf2}{named vector of BF, or matrix of BF with colnames (cols=snps, rows=signals)}

\item{causal.snpid}{Id of the query variant}

\item{pn}{prior probability that none of the SNPS are associated with the queried trait}

\item{pa}{prior probability a SNP other than the causal variant is associated with trait 2}

\item{pc}{prior probability the causal SNP of trait 1 is associated with both traits}
}
\value{
bayes factors of signals
}
\description{
a dataset represented by Bayes factors
}
