% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarsusie.R
\name{cophe.bf_bf}
\alias{cophe.bf_bf}
\title{extract data through Bayes factors}
\usage{
cophe.bf_bf(
  bf2,
  causal.snpid,
  p2a = NULL,
  p12c = NULL,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 5e-06
)
}
\arguments{
\item{bf2}{named vector of BF, or matrix of BF with colnames (cols=snps, rows=signals)}
}
\value{
coloc.signals style result
}
\description{
a dataset represented by Bayes factors
}
\author{
Ichcha Manipur
}
