% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarsusie.R
\name{cophe.bf_bf}
\alias{cophe.bf_bf}
\title{extract data through Bayes factors}
\usage{
cophe.bf_bf(
  sus_dat,
  cred_set,
  querysnpid,
  querytrait,
  pn = NULL,
  pa = NULL,
  pc = NULL,
  ret_pp = TRUE
)
}
\arguments{
\item{sus_dat}{a list with the output of running susie}

\item{cred_set}{credible set extracted from susie}

\item{querysnpid}{Id of the query variant}

\item{querytrait}{Query trait name}

\item{pn}{prior probability that none of the SNPs/variants in the region are associated with the query trait}

\item{pa}{prior probability that a non-query variant is causally associated with the query trait}

\item{pc}{prior probability that the query variant is causally associated with the query trait}
}
\value{
bayes factors of signals
}
\description{
a dataset represented by Bayes factors from SuSIE
}
\seealso{
\code{\link{cophe.susie}}
}
\keyword{internal}
