% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{cophe.single}
\alias{cophe.single}
\title{Bayesian cophescan analysis using Approximate Bayes Factors}
\usage{
cophe.single(
  dataset,
  query.snpid,
  MAF = NULL,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05,
  pa = NULL,
  pc = NULL
)
}
\arguments{
\item{dataset}{a list with specifically named elements defining the dataset
to be analysed.}

\item{query.snpid}{Id of the query variant, (id in dataset$snp)}

\item{MAF}{Minor allele frequency vector}

\item{p1}{prior probability a SNP is associated with trait 1, default 1e-4 (coloc prior)}

\item{p2}{prior probability a SNP is associated with trait 2, default 1e-4 (coloc prior)}

\item{p12}{prior probability a SNP is associated with both traits, default 1e-5 (coloc prior)}

\item{pa}{prior probability that a non-query variant is causally associated with the query trait , default \eqn{pa = p2} (cophescan prior)}

\item{pc}{prior probability that the query variant is causally associated with the query trait, default \eqn{pc =  p12/p1+p12} (cophescan prior)}

\item{pn}{prior probability that none of the SNPs/variants in the region are associated with the query trait , default \eqn{pn = 1 - (pa*(nsnps-1)) - pc} (cophescan prior)}
}
\value{
a list of two \code{data.frame}s:
\itemize{
\item summary is a vector giving the number of SNPs analysed, and the posterior probabilities of Hn (no shared causal variant), Ha (two distinct causal variants) and Hc (one common causal variant)
\item results is an annotated version of the input data containing log Approximate Bayes Factors and intermediate calculations, and the posterior probability SNP.PP.Hc of the SNP being causal for the shared signal \emph{if} Hc is true. This is only relevant if the posterior support for Hc in summary is convincing.
}
}
\description{
Bayesian cophescan analysis under single causal variant assumption
}
\details{
This function calculates posterior probabilities of different
causal variant configurations under the assumption of a single
causal variant for each trait.

If regression coefficients and variances are available, it
calculates Bayes factors for association at each SNP.  If only p
values are available, it uses an approximation that depends on the
SNP's MAF and ignores any uncertainty in imputation.  Regression
coefficients should be used if available. Find more input data structure details
in the coloc package
}
\author{
Ichcha Manipur
}
