% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{cophe_heatmap}
\alias{cophe_heatmap}
\title{Title}
\usage{
cophe_heatmap(multi.dat, thresh = 0.5)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait}

\item{thresh}{Traits with posterior probabilities above this threshold for Hc and Ha will be displayed}
}
\value{
heatmap of posterior probabilities of the phentypes above the set threshold
}
\description{
Title
}
