% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{cophe_heatmap}
\alias{cophe_heatmap}
\title{Heatmap of multi-trait cophescan results}
\usage{
cophe_heatmap(multi.dat, thresh_Hc = 0.5, thresh_Ha = 0.5, ...)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait}

\item{thresh_Hc}{Hc threshold to be displayed}

\item{thresh_Ha}{Ha threshold to be displayed}

\item{...}{additional arguments to be passed to
\link{pheatmap}}
}
\value{
heatmap of posterior probabilities of the phentypes above the set threshold
}
\description{
Heatmap of multi-trait cophescan results
}
