% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{cophe_heatmap}
\alias{cophe_heatmap}
\title{Heatmap of multi-trait cophescan results}
\usage{
cophe_heatmap(
  multi.dat,
  querysnpid,
  query_trait_names,
  thresh_Hc = 0.5,
  thresh_Ha = 0.5,
  ...
)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from \code{cophe.multitrait} or formatted in the same way with multitrait.simplify}

\item{querysnpid}{query variant}

\item{query_trait_names}{names of phenotypes corresponding to the multi.dat results}

\item{thresh_Hc}{Hc threshold to be displayed}

\item{thresh_Ha}{Ha threshold to be displayed}

\item{...}{additional arguments to be passed to
\link[pheatmap]{pheatmap}}
}
\value{
heatmap of posterior probabilities of the phentypes above the set threshold
}
\description{
Heatmap of multi-trait cophescan results
}
