% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{cophe_plot}
\alias{cophe_plot}
\title{cophe_plots showing the Ha and Hc of all traits and labelled above the specified threshold}
\usage{
cophe_plot(
  multi.dat,
  causal.snpid,
  thresh_Hc = 0.5,
  thresh_Ha = 0.5,
  traits.dat = NULL
)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait}

\item{causal.snpid}{query variant}

\item{thresh_Hc}{Hc threshold to be displayed}

\item{thresh_Ha}{Ha threshold to be displayed}

\item{traits.dat}{list of multi-trait oloc structured datasets}
}
\value{
cophescan plots of Ha and Hc
}
\description{
cophe_plots showing the Ha and Hc of all traits and labelled above the specified threshold
}
