% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{cophe_plot}
\alias{cophe_plot}
\title{cophe_plots showing the Ha and Hc of all traits and labelled above the specified threshold}
\usage{
cophe_plot(
  multi.dat,
  query.snpid,
  thresh_Hc = 0.5,
  thresh_Ha = 0.5,
  traits.dat = NULL,
  pheno_names = NULL,
  group_pheno = NULL,
  beta_p = NULL
)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait (Takes row names as the phenotypes if you want an alternative naming pass a vector to pheno_names)}

\item{query.snpid}{query variant}

\item{thresh_Hc}{Hc threshold to be displayed}

\item{thresh_Ha}{Ha threshold to be displayed}

\item{traits.dat}{list of multi-trait oloc structured datasets}

\item{pheno_names}{list of phenotype names}

\item{group_pheno}{Vector with additional grouping of phenotypes}
}
\value{
cophescan plots of Ha and Hc
}
\description{
cophe_plots showing the Ha and Hc of all traits and labelled above the specified threshold
}
