% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophescan-package.R
\docType{package}
\name{cophescan-package}
\alias{cophescan-package}
\alias{cophescan}
\title{The 'cophescan' package.}
\description{
Coloc adapted Phenome-wide Scans
}
\references{
utils::globalVariables(c(".","dfsane","dmvnorm","Hn","Ha","Hc","hit1","hit2","lABF.df2","lABF.h3","lbf1","lbf2","nsnps","snp","snp1","snp2","varbeta","z"))
}
