% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{get_beta}
\alias{get_beta}
\title{Extract beta and p-values of queried variant}
\usage{
get_beta(traits.dat, causal.snpid)
}
\arguments{
\item{traits.dat}{list of cmulti-trait oloc structured datasets}

\item{causal.snpid}{causal.snpid}
}
\value{

}
\description{
Extract beta and p-values of queried variant
}
