% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{get_beta}
\alias{get_beta}
\title{Extract beta and p-values of queried variant}
\usage{
get_beta(traits.dat, querysnpid, querytrait)
}
\arguments{
\item{traits.dat}{list of coloc structured dataset}

\item{querysnpid}{vector of querysnpid}

\item{querytrait}{vector of querytrait names}
}
\value{
data.frame with one column named beta_plot: indicating beta direction (n/p) and another column named pval_plot with -log10(pval) of the queried variant
}
\description{
Extract beta and p-values of queried variant
}
