% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_posterior_prob}
\alias{get_posterior_prob}
\title{Calculation of the posterior prob of Hn, Ha and Hc}
\usage{
get_posterior_prob(params, lbf_mat, nsnps, covar_vec, covar = FALSE)
}
\arguments{
\item{params}{Vector of parameters: \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma}}

\item{lbf_mat}{matrix of log bayes factors: lBF.Ha and lBF.Hc}

\item{nsnps}{number of snps}

\item{covar_vec}{Vector of the covariate}

\item{covar}{logical: should the covariate inflormation be used? default: False}
}
\value{
posterior prob of Hn, Ha and Hc
}
\description{
Calculation of the posterior prob of Hn, Ha and Hc
}
