% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{hypothesis.priors}
\alias{hypothesis.priors}
\title{hypothesis.priors}
\usage{
hypothesis.priors(pn, pa, pc, nsnps)
}
\arguments{
\item{pn}{prior probability none of the SNPS are associated with the queried trait}

\item{pa}{prior probability a SNP other that the causal variant (for a different trait) is associated with the queried trait}

\item{pc}{prior probability that the known causal variant (for a different trait) is associated with the queried trait}

\item{nsnps}{number of SNPs}
}
\value{
hypotheses priors
}
\description{
hypothesis.priors
}
\author{
Ichcha Manipur
}
