% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{hypothesis.priors}
\alias{hypothesis.priors}
\title{hypothesis.priors}
\usage{
hypothesis.priors(nsnps, pn, pa, pc)
}
\arguments{
\item{nsnps}{number of SNPs}

\item{pn}{prior probability that none of the SNPs/variants in the region are associated with the query trait}

\item{pa}{prior probability that a non-query variant is causally associated with the query trait}

\item{pc}{prior probability that the query variant is causally associated with the query trait}
}
\value{
hypotheses priors
}
\description{
Estimate priors for each hypothesis
}
\author{
Ichcha Manipur
}
