% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logd_alpha}
\alias{logd_alpha}
\title{dnorm for alpha}
\usage{
logd_alpha(a, alpha_mean = -10, alpha_sd = 0.5)
}
\arguments{
\item{a}{current alpha}

\item{alpha_mean}{prior for the mean of  alpha}

\item{alpha_sd}{prior for the standard deviation of  alpha}
}
\value{
log dnorm
}
\description{
dnorm for alpha
}
