% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik}
\alias{loglik}
\title{Log likelihood calculation}
\usage{
loglik(params, lbf_mat, nsnps, rg_vec, rg = FALSE)
}
\arguments{
\item{params}{Vector of parameters: α, β and γ}

\item{lbf_mat}{matrix of log bayes factors: lBF.Ha and lBF.Hc}

\item{nsnps}{number of snps}

\item{rg_vec}{Vector of the covariate}

\item{rg}{logical: should the covariate inflormation be used? default: False}
}
\value{
logpost flog of the posteriors
}
\description{
Log likelihood calculation
}
