% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logpost}
\alias{logpost}
\title{Log posterior calculation}
\usage{
logpost(params, lbf_mat, nsnps, covar_vec, covar = FALSE)
}
\arguments{
\item{params}{Vector of parameters: \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma}}

\item{lbf_mat}{matrix of log bayes factors: lBF.Ha and lBF.Hc}

\item{nsnps}{number of snps}

\item{covar_vec}{Vector of the covariate}

\item{covar}{logical: should the covariate inflormation be used? default: False}
}
\value{
logpost flog of the posteriors
}
\description{
Log posterior calculation
}
