% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logpriors}
\alias{logpriors}
\title{Calculate log priors}
\usage{
logpriors(
  params,
  covar = FALSE,
  alpha_mean = -10,
  alpha_sd = 0.5,
  beta_shape = 2,
  beta_scale = 2,
  gamma_shape = 2,
  gamma_scale = 2
)
}
\arguments{
\item{params}{Vector of parameters: \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma}}

\item{covar}{logical: Should the covariate inflormation be used? default: False}

\item{alpha_mean}{prior for the mean of  alpha}

\item{alpha_sd}{prior for the standard deviation of  alpha}

\item{beta_shape}{prior for the shape (gamma distibution) of beta}

\item{beta_scale}{prior for the scale of beta}

\item{gamma_shape}{prior for the shape (gamma distibution) of gamma}

\item{gamma_scale}{prior for the scale of gamma}
}
\value{
log priors
}
\description{
Calculate log priors
}
