% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{multitrait.simplify}
\alias{multitrait.simplify}
\title{Simplifying the output obtained from \code{cophe.multitrait}, \code{cophe.single} or \code{cophe.susie}}
\usage{
multitrait.simplify(multi.dat, only_BF = FALSE)
}
\arguments{
\item{multi.dat}{output obtained from \code{cophe.multitrait}, \code{cophe.single} or \code{cophe.susie}}

\item{only_BF}{return only bayes factors and not posterior probabilities (default=FALSE)}
}
\value{
dataframe with posterior probabilties of Hn, Hc and Ha
}
\description{
Simplifying the output obtained from \code{cophe.multitrait}, \code{cophe.single} or \code{cophe.susie}
}
