% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pars2pik}
\alias{pars2pik}
\title{Conversion of parameters alpha, beta and gamma to pnk, pak and pck}
\usage{
pars2pik(params, nsnps, covar_vec, covar = FALSE)
}
\arguments{
\item{params}{Vector of parameters: \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma}}

\item{nsnps}{number of snps}

\item{covar_vec}{Vector of the covariate}

\item{covar}{logical: should the covariate information be used? default: False}
}
\value{
pik matrix of priors: pnk, pak and pck
}
\description{
Conversion of parameters alpha, beta and gamma to pnk, pak and pck
}
