% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{per.snp.priors}
\alias{per.snp.priors}
\title{per.snp.priors}
\usage{
per.snp.priors(
  nsnps,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05,
  pa = NULL,
  pc = NULL
)
}
\arguments{
\item{nsnps}{number of SNPs}

\item{p1}{prior probability a SNP is associated with trait 1, default 1e-4}

\item{p2}{prior probability a SNP is associated with trait 2, default 1e-4}

\item{p12}{prior probability a SNP is associated with both traits, default 1e-5}

\item{pa}{prior probability a SNP other that the causal variant (for a different trait) is associated with the queried trait , default \eqn{pa = p2}}

\item{pc}{prior probability that the known causal variant (for a different trait) is associated with the queried trait, default \eqn{pc =  p12/p1+p12}}

\item{pn}{prior probability that none of the SNPS are associated with the queried trait , default \eqn{pn = 1 - (pa*(nsnps-1)) - pc}}
}
\value{
priors at the causal variant
}
\description{
Estimate per snp priors
}
\author{
Ichcha Manipur
}
