% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piks}
\alias{piks}
\title{List of posterior probabilities: Hn, Ha and Hc over all iterations}
\usage{
piks(params, nsnps, rg_vec, rg = FALSE)
}
\arguments{
\item{params}{Vector of parameters: α, β and γ}

\item{nsnps}{number of snps}

\item{rg_vec}{Vector of the covariate}

\item{rg}{logical: was the covariate inflormation  used? default: False}
}
\value{
List of priors (len: iterations): p0k, pak and pck
}
\description{
List of posterior probabilities: Hn, Ha and Hc over all iterations
}
