% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piks}
\alias{piks}
\title{List of priors: pn, pa and pc over all iterations}
\usage{
piks(params, nsnps, covar_vec, covar = FALSE)
}
\arguments{
\item{params}{Vector of parameters: \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma}}

\item{nsnps}{number of snps}

\item{covar_vec}{Vector of the covariate}

\item{covar}{logical: was the covariate inflormation  used? default: False}
}
\value{
List of priors (len: iterations): pnk, pak and pck
}
\description{
List of priors: pn, pa and pc over all iterations
}
