% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{plot_trait_manhat}
\alias{plot_trait_manhat}
\title{Plot region Manhattan for a trait highlighting the queried variant}
\usage{
plot_trait_manhat(trait.dat, querysnpid, alt.snpid = NULL)
}
\arguments{
\item{trait.dat}{dataset used as input for running cophescan}

\item{querysnpid}{the id of the causal variant as present in trait.dat$snp, plotted in red}

\item{alt.snpid}{the id of the other variants as a vector to be plotted, plotted in blue}
}
\value{
regional manhattan plot
}
\description{
Plot region Manhattan for a trait highlighting the queried variant
}
