% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posterior_prob}
\alias{posterior_prob}
\title{List of posterior probabilities: Hn, Ha and Hc over all iterations}
\usage{
posterior_prob(params, lbf_mat, nsnps, rg_vec, rg = FALSE)
}
\arguments{
\item{params}{Vector of parameters: α, β and γ}

\item{lbf_mat}{matrix of log bayes factors: lBF.Ha and lBF.Hc}

\item{nsnps}{number of snps}

\item{rg_vec}{Vector of the covariate}

\item{rg}{logical: was the covariate inflormation  used? default: False}
}
\value{
params List of posterior probabilties (len: iterations): Hn, Ha and Hc
}
\description{
List of posterior probabilities: Hn, Ha and Hc over all iterations
}
