% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posterior_prob}
\alias{posterior_prob}
\title{List of posterior probabilities: Hn, Ha and Hc over all iterations}
\usage{
posterior_prob(params, lbf_mat, nsnps, covar_vec, covar = FALSE)
}
\arguments{
\item{params}{Vector of parameters: \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma}}

\item{lbf_mat}{matrix of log bayes factors: lBF.Ha and lBF.Hc}

\item{nsnps}{number of snps}

\item{covar_vec}{Vector of the covariate}

\item{covar}{logical: was the covariate inflormation  used? default: False}
}
\value{
List of posterior probabilties (len: iterations): Hn, Ha and Hc
}
\description{
List of posterior probabilities: Hn, Ha and Hc over all iterations
}
