% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copheplots.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare data for plotting}
\usage{
prepare_plot_data(
  multi.dat,
  querysnpid,
  query_trait_names,
  thresh_Ha = 0.5,
  thresh_Hc = 0.5,
  hmp = FALSE,
  cophe.plot = TRUE
)
}
\arguments{
\item{multi.dat}{multi trait cophescan results returned from cophe.multitrait or multitrait.simplify}

\item{querysnpid}{query variant}

\item{query_trait_names}{vector of names of the query traits}

\item{thresh_Ha}{Ha threshold to be displayed}

\item{thresh_Hc}{Hc threshold to be displayed}

\item{hmp}{return for heatmap}

\item{cophe.plot}{default: TRUE, return for \code{cophe_plot}}
}
\value{
plot list
}
\description{
Prepare data for plotting
}
\seealso{
\code{\link{cophe_plot}}, \code{\link{cophe.susie}}, \code{\link{cophe.multitrait}},  \code{\link{multitrait.simplify}}
default NULL
}
