% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrop_hier_priors.R
\name{run_metrop_priors}
\alias{run_metrop_priors}
\title{run_metrop_priors}
\usage{
run_metrop_priors(
  multi.dat,
  rg = FALSE,
  rg_vec = NULL,
  nits = 10000,
  thin = 1,
  posterior = F,
  avg_posterior = F,
  pik = T,
  chains = 1,
  cores = 1
)
}
\arguments{
\item{multi.dat}{matrix of bf values}

\item{rg}{whether to run rg}

\item{rg_vec}{vector of genetic correlations}

\item{nits}{number of iterations}

\item{thin}{burnin}

\item{posterior}{default: F, estimate posterior probabilities of the hypotheses}

\item{avg_posterior}{default: F, estimate the average of the posterior probabilities of the hypotheses}

\item{pik}{default: F, inferred prior probabilities}

\item{chains}{default: 1, number of chains}

\item{cores}{default: 1, number of cores}
}
\value{
list containing posterior of the parameters
}
\description{
run_metrop_priors
}
