% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrop_hier_priors.R
\name{run_metrop_priors}
\alias{run_metrop_priors}
\title{Run the hierarchical model to infer priors}
\usage{
run_metrop_priors(
  multi.dat,
  rg = FALSE,
  rg_vec = NULL,
  nits = 10000,
  thin = 1,
  posterior = F,
  avg_pik = T,
  avg_posterior = T,
  pik = F,
  chains = 1,
  cores = 1,
  alpha_mean = -10,
  alpha_sd = 0.5,
  beta_shape = 2,
  beta_scale = 2,
  gamma_shape = 2,
  gamma_scale = 2
)
}
\arguments{
\item{multi.dat}{matrix of bf values}

\item{rg}{whether to run rg}

\item{rg_vec}{vector of genetic correlations}

\item{nits}{number of iterations}

\item{thin}{burnin}

\item{posterior}{default: F, estimate posterior probabilities of the hypotheses}

\item{avg_pik}{default: F, estimate the average of the pik}

\item{avg_posterior}{default: F, estimate the average of the posterior probabilities of the hypotheses}

\item{pik}{default: F, inferred prior probabilities}

\item{chains}{default: 1, number of chains}

\item{cores}{default: 1, number of cores}

\item{alpha_mean}{prior for the mean of  alpha}

\item{alpha_sd}{prior for the standard deviation of  alpha}

\item{beta_shape}{prior for the shape (gamma distibution) of beta}

\item{beta_scale}{prior for the scale of beta}

\item{gamma_shape}{prior for the shape (gamma distibution) of gamma}

\item{gamma_scale}{prior for the scale of gamma}
}
\value{
matrix with average of all the posterior probabilities: Hn, Ha and Hc

list containing posterior of the parameters
}
\description{
Run the hierarchical model to infer priors
}
