% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_gamma}
\alias{sample_gamma}
\title{sample gamma}
\usage{
sample_gamma(gamma_shape = 2, gamma_scale = 2)
}
\arguments{
\item{gamma_shape}{prior for the shape (gamma distibution) of gamma}

\item{gamma_scale}{prior for the scale of gamma}
}
\value{
sample from rgamma for \eqn{\gamma}
}
\description{
sample gamma
}
