% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{summary.cophe}
\alias{summary.cophe}
\title{print the summary of results from cophescan single or susie}
\usage{
\method{summary}{cophe}(cophe.res, ...)
}
\arguments{
\item{cophe.res}{Result from either cophe.susie or cophe.single}
}
\value{
log bayes and posterior probabilities
}
\description{
print the summary of results from cophescan single or susie
}
