% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlevar.R
\name{summary.cophe}
\alias{summary.cophe}
\title{print the summary of results from cophescan single or susie}
\usage{
\method{summary}{cophe}(object, ...)
}
\arguments{
\item{object}{Result from either \code{cophe.susie} or \code{cophe.single}}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
log bayes and posterior probabilities
}
\description{
print the summary of results from cophescan single or susie
}
\seealso{
\code{\link{cophe.single}}, \code{\link{cophe.susie}}
}
